v0.3 1407 Contact Group Edit

+++ app/Http/Controllers/GroupsController.php
@@ -4,8 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Group;
 use App\Customer;
+use App\Contact;
+use App\Contactgroupcombo;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Session;
 
 class GroupsController extends Controller
 {
@@ -33,12 +37,28 @@ class GroupsController extends Controller
         return redirect('/groups')->with('success', 'Group added');
     }
 
+    public function edit($id)
+    {
+        $group = Group::where('customer_id', '=', '1')->find($id);
+        $membership = Contactgroupcombo::where('customer_id', '=', '1')->select('contact_id')->where('group_id', '=', $id)->get();
+        if ($group == null){
+            return redirect('/groups')->with('error', 'Group not found');
+        }
+
+        return view('groups.edit',['group' => $group, 'membership' => $membership]);
+    }
+
     public function update(Request $request, $id)
     {
         $group = Group::where('id', '=', $id)->first();
-        $this->validate($request, [
+        $validator = Validator::make($request->all(), [
             'name' => 'required'
         ]);
+        if ($validator->fails())
+        {
+            $error = $validator->errors()->all();
+            return redirect()->back()->with('error', "Please enter the required fields");
+        }
         
         $user_id = Auth::user()->id;
 
@@ -56,6 +76,12 @@ class GroupsController extends Controller
 
     public function destroy($id)
     {
+        $contactgroupcombo = Contactgroupcombo::where('customer_id', '=', '1')->select('id')->where('group_id', '=', $id)->pluck('id')->toArray();
+
+        if (!empty($contactgroupcombo)) 
+        {
+            return redirect('/groups')->with('error', 'Disassociate contacts before you delete the group');
+        }
         $group = Group::find($id);
         $group->delete();
         return redirect('/groups')->with('success', 'Group Deleted');
